requirements.txt
openai-agents==0.0.4
python-dotenv==1.0.1
aiogram == 3.20.0